Moved 5 funcs to `consequences.py`: create_new_cat_block, create_new_cat, gather_cat_objects, unpack_rel_block, change_relationship_values
