+++ scripts/events_module/relationship/pregnancy_events.py
@@ -15,10 +15,12 @@ from scripts.game_structure import constants
 from scripts.game_structure import game
 from scripts.game_structure.localization import load_lang_resource
 from scripts.utility import (
-    create_new_cat,
     event_text_adjust,
-    change_relationship_values,
     adjust_list_text,
+)
+from scripts.events_module.consequences import (
+    create_new_cat,
+    change_relationship_values,
 )
 from scripts.events_module.event_filters import (
     get_highest_romantic_relation,
