+++ scripts/events_module/relationship/welcoming_events.py
@@ -11,9 +11,9 @@ from scripts.cat.enums import CatRank
 from scripts.event_class import Single_Event
 from scripts.game_structure import game
 from scripts.utility import (
-    change_relationship_values,
     event_text_adjust,
 )
+from scripts.events_module.consequences import change_relationship_values
 from scripts.game_structure.localization import load_lang_resource
